% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVars.R
\name{addVars}
\alias{addVars}
\title{addVars}
\usage{
addVars(match, data, ...)
}
\arguments{
\item{match}{list structure with possibly matched observations}

\item{data}{list of data frames}

\item{...}{named list of variables (one for each data frame)}
}
\value{
list structure with possibly matched observations with added variables
}
\description{
Adds further variables for inspection to the match without any check.
}
\examples{
set.seed(0)
# create two data sets which consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
match <- findMatch(x, c('code', 'code'))
match <- addVars(match, x, birthday=c('birthday', 'birthday'))
head(match)
summary(match)
}
