% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDups.R
\name{deleteDups}
\alias{deleteDups}
\title{deleteDups}
\usage{
deleteDups(match)
}
\arguments{
\item{match}{match structure}
}
\value{
a modified match
}
\description{
Deletes duplicate entries according to some variable in the match
}
\examples{
set.seed(0)
# create two data sets which consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
match <- findMatch(x, c('code', 'code'))
summary(match)
match <- deleteDups(match)
summary(match)
}
