% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsVar.R
\name{existsVars}
\alias{existsVars}
\title{existsVars}
\usage{
existsVars(name, data)
}
\arguments{
\item{name}{chracter: names of variables}

\item{data}{data frame}
}
\value{
\code{TRUE} if all names in \code{name} exists in \code{data} otherwise \code{FALSE}
}
\description{
Checks if all variables in \code{name} exists in the data frame \code{data}.
}
