% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDuplicates.R
\name{findDuplicates}
\alias{findDuplicates}
\title{findDuplicates}
\usage{
findDuplicates(data, var, dmax = 3, exclude = c("", "."),
  ignore.case = FALSE)
}
\arguments{
\item{data}{data frame}

\item{var}{character: name of variable}

\item{dmax}{maximal levensthein distance for matching in text variables $l(t_{i1},t{j2]}<dmax$), defaults to \code{3}}

\item{exclude}{entries to be excluded from the unique values, defaults to \code{c('', '.')}}

\item{ignore.case}{if FALSE, the uniques values are case sensitive and if TRUE, case is ignored}
}
\value{
a list structure with possibly duplicates
}
\description{
Finds possible duplicates in data set.
}
\examples{
# create a data set
x1 <- generateTestData(500)
#
match <- findDuplicates(x1, 'code')
head(match)
}
