% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMatch.R
\name{findMatch}
\alias{findMatch}
\alias{findMatch.default}
\title{findMatch}
\usage{
findMatch(data, ...)

\method{findMatch}{default}(data, vars, dmax = 3, exclude = c("", "."),
  ignore.case = FALSE, unique.id = NULL, output = 50,
  cmpfunc = NULL, ...)
}
\arguments{
\item{data}{list of data frames}

\item{...}{further parameters for cmp}

\item{vars}{vector of variables. One for each data frame.}

\item{dmax}{maximal levensthein distance for matching in text variables $l(t_{i1},t{j2]}<dmax$), defaults to \code{3}}

\item{exclude}{entries to be excluded from the unique values, defaults to \code{c('', '.')}}

\item{ignore.case}{if FALSE, the uniques values are case sensitive and if TRUE, case is ignored}

\item{unique.id}{vector of variables which contain a unique ID over all data sets. If not given then \code{filename:lineno} will be used.}

\item{output}{number of observation to analyse before a progress information is displayed}

\item{cmpfunc}{function for comparison of strings of form \code{fun(x, y, ignore.case, ...)} (default: \code{\link[utils]{adist}})}
}
\value{
a list structure with possibly matched observations

a list structure with possibly matched observations
}
\description{
Finds matches between two or more data sets based on a text variable (code or e-mail) based on
\href{https://en.wikipedia.org/wiki/Levenshtein_distance}{Levensthein distances}. For a detailed application see the vignette.
}
\details{
The result consists of a list with three elements
\describe{
\item{\code{line}}{a matrix with the line numbers of the matching observations}
\item{\code{idn}}{a matrix with the common ID \code{ZDV} and the original text variables in the data sets}
\item{\code{leven}}{a matrix with the levenshtein distance between the common ID and the original text variables in the data sets}
}
}
\examples{
set.seed(0)
# create two data sets where the second consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
# match by code
match <- findMatch(x, c('code', 'code'))
head(match)
summary(match)
}
