% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPanel.R
\name{findPanel}
\alias{findPanel}
\title{findPanel}
\usage{
findPanel(data, matches)
}
\arguments{
\item{data}{list of data frames}

\item{matches}{a double list of matches with elements \code{t1}, \code{t2}, ...}
}
\value{
matches
}
\description{
findPanel
}
