% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTestData.R
\name{generateTestData}
\alias{generateTestData}
\title{generateTestData}
\usage{
generateTestData(n = 500, x = NULL, mortality = 0.5, ...)
}
\arguments{
\item{n}{number of observations for the test data (default: 500)}

\item{x}{data set to take observations from for simulating longitudinal data}

\item{mortality}{mortality rate (default: 0.5)}

\item{...}{further \strong{named} (artificial) variables to add to the data set}
}
\value{
a data frame
}
\description{
Generates randomly a test data set of size \code{n}.
}
\details{
The generated data set contains at least eight variables:
\describe{
\item{\code{sex}}{sex of respondent (\code{M} male, \code{F} female)}
\item{\code{firstname}}{first name of respondent}
\item{\code{mothername}}{first name of respondents mother}
\item{\code{lastname}}{last name of respondent}
\item{\code{email}}{e-mail address of respondent}
\item{\code{birthplace}}{birthplace of respondent}
\item{\code{birthday}}{birthday of respondent}
\item{\code{code}}{code generated from respondents data: 2nd and 3rd letter of mothers first name, day of birth, 3rd and 4th letter of birth city and 1st letter of first name}
}

The basic data files for the generation can be found in the \code{extdata} directory of the package
\describe{
\item{\code{firstname_female.txt}}{taken from \href{https://de.wiktionary.org/wiki/Verzeichnis:International/Weibliche_Vornamen}{Wiktionary}}
\item{\code{firstname_male.txt}}{taken from \href{https://de.wiktionary.org/wiki/Verzeichnis:International/M\%C3\%A4nnliche_Vornamen}{Wiktionary}}
\item{\code{lastname.txt}}{taken from \href{http://wiki-de.genealogy.net/Die_1000_h\%C3\%A4ufigsten_Familiennamen_in_Deutschland}{genealogy.net Wiki}}
\item{\code{email_provider.txt}}{taken from \href{https://gist.github.com/tbrianjones/5992856}{T. Brian Jones \code{free_email_provider_domains.txt}}}
\item{\code{staedte.txt}}{taken from \href{https://de.wikipedia.org/wiki/Liste_der_St\%C3\%A4dte_in_Deutschland}{Wikipedia list of cities and towns in Germany}}
}
}
\examples{
# create two data sets where the second consists of 50\% observations of the first
n1 <- 500
n2 <- 300
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
x1$points <- findInterval(rnorm(n1, mean=10, sd=3), 0:15)-1
head(x1)
x2$points <- findInterval(rnorm(n2, mean=10, sd=3), 0:15)-1
head(x2)
}
