% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.findMatch.R
\name{head.findMatch}
\alias{head.findMatch}
\alias{tail.findMatch}
\title{head/tail}
\usage{
\method{head}{findMatch}(x, n = 6L, ...)

\method{tail}{findMatch}(x, n = 6L, ...)
}
\arguments{
\item{x}{a list structure with possibly matched observations}

\item{n}{number of lines to show, defaults to 6}

\item{...}{further arguments (unused)}
}
\value{
the first/last n possible matches as data frame
}
\description{
\code{head} shows the first \code{n} possible matches and \code{tail} shows the last \code{n} possible matches.
}
\examples{
set.seed(0)
# create two data sets where the second consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
#
match <- findDuplicates(x[[1]], 'code')
head(match)
}
