% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMerge.R
\name{makeMerge}
\alias{makeMerge}
\title{makeMerge}
\usage{
makeMerge(data, files, full = FALSE, duplicates.ok = FALSE, ...)
}
\arguments{
\item{data}{list of data frames}

\item{files}{list of files which contain the actual matches}

\item{full}{logical: return full data set (non-matching are observations  are filels with NA) or just matched obs, defaults to \code{FALSE}}

\item{duplicates.ok}{logical: Are duplicates ok, defaults ton \code{FALSE}}

\item{...}{further parameters given to \code{\link[rio]{import}} for importing match data}
}
\value{
a data frame with merged data
}
\description{
Creates a a data set consisting of all variables according to the matching results
}
\examples{
set.seed(0)
# create two data sets where the second consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
match <- findMatch(x, c('code', 'code'))
head(match)
\dontrun{
exportMatch(match, 'match.xlsx')
# creates an error since we have duplicate matches
data <- makeMerge(x, 'match.xlsx')
}
}
