% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numIncrease.R
\name{numIncrease}
\alias{numIncrease}
\title{numIncrease}
\usage{
numIncrease(match, data, min = rep(0, length(data)),
  max = 0:(length(data) - 1), ...)
}
\arguments{
\item{match}{match structure}

\item{data}{list of data frames}

\item{min}{minimal in/decrease, defaults to \code{rep(0, length(data))}}

\item{max}{maximal in/decrease, defaults to \code{0:(length(data)-1)}}

\item{...}{named list of variables (one for each data frame)}
}
\value{
updated match structure
}
\description{
Checks for two possible matches the intervals $\[t_{i,j}-max_j, t_{i,j}-min_j\]$ intersect.
If the condition fails than each Levenshtein distance will be increased by one.
}
\examples{
set.seed(0)
# create two data sets where the second consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
# create ages in years from birthdays
today  <- as.Date(Sys.time())
x[[1]]$age <- as.numeric(trunc(difftime(today, x[[1]]$birthday, unit="days")/365))
x[[2]]$age <- as.numeric(trunc(difftime(today+365, x[[2]]$birthday, unit="days")/365))
#
match <- findMatch(x, c('code', 'code'))
summary(match)
match <- numIncrease(match, x, age=c('age', 'age'))
summary(match)
head(match)
\dontrun{
# with \%>\% operator
library('magrittr')
match <- findMatch(x, c('code', 'code')) \%>\%
         numIncrease(x, age=c('age', 'age'))
}
}
