% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.findMatch.R
\name{summary.findMatch}
\alias{summary.findMatch}
\title{summary.findMacth}
\usage{
\method{summary}{findMatch}(object, ...)
}
\arguments{
\item{object}{a list structure with possibly matched observations}

\item{...}{further parameters (unused)}
}
\value{
summary of possible matches
}
\description{
Shows the numer of possible matches and, if any matches available, a contingency table of
}
\examples{
# create two data sets where the second consists of 50\% observations of the first
n1 <- n2 <- 500
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
x1$points <- findInterval(rnorm(n1, mean=10, sd=3), 0:15)-1
x2$points <- findInterval(rnorm(n2, mean=10, sd=3), 0:15)-1
#
match <- findMatch(list(x1,x2), c('code', 'code'))
head(match)
summary(match)
}
