% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(x)
}
\arguments{
\item{x}{character vector}
}
\value{
cleaned character vector
}
\description{
Deletes leading and trailing whitespaces in a string.
}
\examples{
trim (' Hello World ')
}
\references{
\href{https://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace}{Stack overflow: How to trim leading and trailing whitespace?}
}
