% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVars.R
\name{addVars}
\alias{addVars}
\title{addVars}
\usage{
addVars(match, data, ...)
}
\arguments{
\item{match}{list structure with possibly matched observations}

\item{data}{list of data frames}

\item{...}{named list of variables (one for each data frame)}
}
\value{
list structure with possibly matched observations with added variables
}
\description{
Adds further variables for inspection to the match without any check.
}
\examples{
#' # create two data sets where the second consists of 50\% observations of the first
n1 <- n2 <- 500
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
x1$points <- findInterval(rnorm(n1, mean=10, sd=3), 0:15)-1
x2$points <- findInterval(rnorm(n2, mean=10, sd=3), 0:15)-1
#
match <- findMatch(list(x1,x2), c('code', 'code'))
match <- addVars(match, list(x1,x2), birthday=c('birthday', 'birthday'))
head(match)
summary(match)
}
