% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allIdentical.R
\name{allIdentical}
\alias{allIdentical}
\title{allIdentical}
\usage{
allIdentical(match, data, vars, name = NULL)
}
\arguments{
\item{match}{match structure}

\item{data}{list of data frames}

\item{vars}{vector of variables. One for each data frame.}

\item{name}{list element name to store original variables}
}
\description{
Increases each Levenshtein distances by one if variable entries do not match
}
\examples{
# create two data sets where the second consists of 50\% observations of the first
n1 <- n2 <- 500
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
x1$points <- findInterval(rnorm(n1, mean=10, sd=3), 0:15)-1
x2$points <- findInterval(rnorm(n2, mean=10, sd=3), 0:15)-1
#
match <- findMatch(list(x1,x2), c('code', 'code'))
summary(match)
match <- allIdentical(match, list(x1,x2), c('sex', 'sex'), 'sex')
summary(match)
\dontrun{
# with \%>\% operator
library('magrittr')
match <- findMatch(list(x1,x2), c('code', 'code')) \%>\%
         allIdentical(list(x1,x2), c('sex', 'sex'), 'sex')
}
}
