% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allIdentical.R
\name{allIdentical}
\alias{allIdentical}
\title{allIdentical}
\usage{
allIdentical(match, data, ...)
}
\arguments{
\item{match}{match structure}

\item{data}{list of data frames}

\item{...}{named list of variables (one for each data frame)}
}
\value{
returns a modified match
}
\description{
Increases each Levenshtein distances by one if variable entries do not match
}
\examples{
set.seed(0)
# create two data sets which consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
match <- findMatch(x, c('code', 'code'))
summary(match)
match <- allIdentical(match, x, sex=c('sex', 'sex'))
summary(match)
\dontrun{
# with \%>\% operator
library('magrittr')
match <- findMatch(x, c('code', 'code')) \%>\%
         allIdentical(x, sex=c('sex', 'sex'))
}
}
