% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDups.R
\name{deleteDups}
\alias{deleteDups}
\title{deleteDups}
\usage{
deleteDups(match)
}
\arguments{
\item{match}{match structure}
}
\value{
a modified match
}
\description{
Deletes duplicate entries according to some variable in the match
}
