% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsVar.R
\name{existsVars}
\alias{existsVars}
\title{existsVars}
\usage{
existsVars(name, data)
}
\arguments{
\item{name}{chracter: names of variables}

\item{data}{data frame}
}
\value{
\code{TRUE} if all names in \code{name} exists in \code{data} otherwise \code{FALSE}
}
\description{
Checks if all variables in \code{name} exists in the data frame \code{data}.
}
\examples{
#' set.seed(0)
# create two data sets which consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
existsVars("Geburtstag", x[[1]])
existsVars("birthday", x[[2]])
}
