% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMatch.R
\name{exportMatch}
\alias{exportMatch}
\title{exportMatch}
\usage{
exportMatch(match, file, ...)
}
\arguments{
\item{match}{match structure}

\item{file}{file name}

\item{...}{further parameters given to \code{\link[rio]{export}} method oth the library \code{rio}}
}
\value{
a file
}
\description{
Exports the found matches into a data file for further manual inspection.
Sort the matched observations by their respective levensthein distance and delete non-matches from the file.
}
\examples{
set.seed(0)
library("findMatch")
# create two data sets where the second consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
match <- findMatch(x, c('code', 'code'))
exportMatch(match, 'match.xlsx', overwrite=TRUE)
}
