% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMatch.R
\name{findMatch}
\alias{findMatch}
\alias{findMatch.default}
\title{findMatch}
\usage{
findMatch(data, ...)

\method{findMatch}{default}(data, vars, dmax = 3, exclude = c("", "."),
  ignore.case = FALSE, ...)
}
\arguments{
\item{data}{list of data frames}

\item{...}{further parameters}

\item{vars}{vector of variables. One for each data frame.}

\item{dmax}{maximal levensthein distance for matching in text variables $l(t_{i1},t{j2]}<dmax$), defaults to \code{3}}

\item{exclude}{entries to be excluded from the unique values, defaults to \code{c('', '.')}}

\item{ignore.case}{if FALSE, the uniques values are case sensitive and if TRUE, case is ignored}
}
\value{
a list structure with possibly matched observations

a list structure with possibly matched observations
}
\description{
Finds matches between two or more data sets based on a text variable (code or e-mail) based on
\href{https://en.wikipedia.org/wiki/Levenshtein_distance}{Levensthein distances}.
}
\details{
The result consists of a list with three elements
\describe{
\item{\code{line}}{a matrix with the line numbers of the matching observations}
\item{\code{idn}}{a matrix with the common ID \code{ZDV} and the original text variables in the data sets}
\item{\code{leven}}{a matrix with the levenshtein distance between the common ID and the original text variables in the data sets}
}
}
\examples{
# create two data sets where the second consists of 50\% observations of the first
n1 <- n2 <- 500
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
x1$points <- findInterval(rnorm(n1, mean=10, sd=3), 0:15)-1
x2$points <- findInterval(rnorm(n2, mean=10, sd=3), 0:15)-1
#
match <- findMatch(list(x1,x2), c('code', 'code'))
head(match)
summary(match)
}
