% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.findMatch.R
\name{head.findMatch}
\alias{head.findMatch}
\alias{tail.findMatch}
\title{head/tail}
\usage{
\method{head}{findMatch}(x, n = 6L, ...)

\method{tail}{findMatch}(x, n = 6L, ...)
}
\arguments{
\item{x}{a list structure with possibly matched observations}

\item{n}{number of lines to show, defaults to 6}

\item{...}{further arguments (unused)}
}
\value{
the first/last n possible matches as data frame
}
\description{
\code{head} shows the first \code{n} possible matches and \code{tail} shows the last \code{n} possible matches.
}
\examples{
# create a data set
x1 <- generateTestData(500)
#
match <- findDuplicates(x1, 'code')
head(match)
}
