% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMerge.R
\name{makeMerge}
\alias{makeMerge}
\title{makeMerge}
\usage{
makeMerge(data, files, full = FALSE, duplicates.ok = FALSE, ...)
}
\arguments{
\item{data}{list of data frames}

\item{files}{list of files which contain the actual matches}

\item{full}{logical: return full data set (non-matching are observations  are filels with NA) or just matched obs, defaults to \code{FALSE}}

\item{duplicates.ok}{logical: Are duplicates ok, defaults ton \code{FALSE}}

\item{...}{further parameters given to \code{\link[rio]{import}} for importing match data}
}
\value{
a data frame with merged data
}
\description{
Creates a a data set consisting of all variables according to the matching results
}
\examples{
# create two data sets where the second consists of 50\% observations of the first
n1 <- n2 <- 500
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
#
match <- findMatch(list(x1,x2), c('code', 'code'))
head(match)
\dontrun{
exportMatch(match, 'match.xlsx')
# creates an error since we have duplicate matches
data <- makeMerge(list(x1, x2), 'match.xlsx')
}
}
