% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numIncrease.R
\name{numIncrease}
\alias{numIncrease}
\title{numIncrease}
\usage{
numIncrease(match, data, min = rep(0, length(data)),
  max = 0:(length(data) - 1), ...)
}
\arguments{
\item{match}{match structure}

\item{data}{list of data frames}

\item{min}{minimal in/decrease, defaults to \code{rep(0, length(data))}}

\item{max}{maximal in/decrease, defaults to \code{0:(length(data)-1)}}

\item{...}{named list of variables (one for each data frame)}
}
\value{
updated match structure
}
\description{
Checks for two possible matches the intervals $\[t_{i,j}-max_j, t_{i,j}-min_j\]$ intersect.
If the condition fails than each Levenshtein distance will be increased by one.
}
\examples{
# create two data sets where the second consists of 50\% observations of the first
n1 <- n2 <- 500
x1 <- generateTestData(n1)
x2 <- generateTestData(n2, x1)
# create ages in years from birthdays
today  <- as.Date(Sys.time())
x1$age <- as.numeric(trunc(difftime(today, x1$birthday, unit="days")/365))
x2$age <- as.numeric(trunc(difftime(today+365, x2$birthday, unit="days")/365))
#
match <- findMatch(list(x1,x2), c('code', 'code'))
summary(match)
match <- numIncrease(match, list(x1,x2), age=c('age', 'age'))
summary(match)
head(match)
\dontrun{
# with \%>\% operator
library('magrittr')
match <- findMatch(list(x1,x2), c('code', 'code')) \%>\%
         numIncrease(list(x1,x2), age=c('age', 'age'), 'age')
}
}
