% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.findMatch.R
\name{summary.findMatch}
\alias{summary.findMatch}
\title{summary.findMacth}
\usage{
\method{summary}{findMatch}(object, ...)
}
\arguments{
\item{object}{a list structure with possibly matched observations}

\item{...}{further parameters (unused)}
}
\value{
summary of possible matches
}
\description{
Shows the numer of possible matches and, if any matches available, a contingency table of
}
\examples{
set.seed(0)
# create two data sets where the second consists of
# 200 obs. only in t1, 200 obs. in t1 and t2 and
# 100 obs. only in t2
n <- list(c(200, 1), c(200, 1, 2), c(100, 2))
x <- generateTestData(n)
#
match <- findMatch(x, c('code', 'code'))
head(match)
summary(match)
}
